refactor: move `Release` classes under `Version`
These all deal with releasing a specific **version**, so they belong
under that namespace.

This applies to the command name as well, which now becomes
`version:release`.
