+++ README.md
@@ -277,8 +277,6 @@ The default providers available are:
 
 ## Supported commands
 
-- `release`: create a new release on your given provider, using the relevant
-  changelog entry. This command also pushes the related tag.
 - `bump`, `bump:bugfix`, and `bump:patch`: Create a new changelog entry for the
   next bugfix release.
 - `bump:minor`: Create a new changelog entry for the next minor release.
@@ -298,6 +296,8 @@ The default providers available are:
 - `entry:fixed`: Create a changelog entry in the "Fixed" section.
 - `version:edit`: Edit a full changelog version and its entries.
 - `version:list`: List all changelog versions currently in the file.
+- `version:release`: create a new release on your given provider, using the
+  relevant changelog entry. This command also pushes the related tag.
 - `version:remove`: Remove the given changelog version and its entries.
 - `version:show`: Show the given changelog version and its entries.
 - `version:tag`: Create a new tag, using the relevant changelog entry. Creates
