+++ bin/keep-a-changelog
@@ -54,9 +54,9 @@ $application->addCommands([
     new Entry\EntryCommand($dispatcher, 'entry:fixed'),
     new Changelog\NewCommand($dispatcher, 'changelog:new'),
     new Changelog\ReadyCommand($dispatcher, 'changelog:ready'),
-    new Release\ReleaseCommand($dispatcher, 'release'),
     new Version\EditCommand($dispatcher, 'version:edit'),
     new Version\ListCommand($dispatcher, 'version:list'),
+    new Version\ReleaseCommand($dispatcher, 'version:release'),
     new Version\RemoveCommand($dispatcher, 'version:remove'),
     new Version\ShowCommand($dispatcher, 'version:show'),
     new Version\TagCommand($dispatcher, 'version:tag'),
