+++ src/ListenerProvider.php
@@ -77,17 +77,6 @@ class ListenerProvider implements ListenerProviderInterface
             Entry\PrependPatchLinkListener::class,
             Entry\AddChangelogEntryListener::class,
         ],
-        Release\ReleaseEvent::class                => [
-            Release\ConfigListener::class,
-            Release\VerifyTagExistsListener::class,
-            Release\VerifyProviderCanReleaseListener::class,
-            Common\IsChangelogReadableListener::class,
-            Common\ParseChangelogListener::class,
-            Common\FormatChangelogListener::class,
-            Release\PushTagToRemoteListener::class,
-            Release\CreateReleaseNameListener::class,
-            Release\PushReleaseToProviderListener::class,
-        ],
         Version\EditChangelogVersionEvent::class   => [
             Config\ConfigListener::class,
             Common\IsChangelogReadableListener::class,
@@ -100,6 +89,17 @@ class ListenerProvider implements ListenerProviderInterface
             Config\ConfigListener::class,
             Common\IsChangelogReadableListener::class,
             Version\ListVersionsListener::class,
+        ],
+        Version\ReleaseEvent::class                => [
+            Version\ReleaseCommandConfigListener::class,
+            Version\VerifyTagExistsListener::class,
+            Version\VerifyProviderCanReleaseListener::class,
+            Common\IsChangelogReadableListener::class,
+            Common\ParseChangelogListener::class,
+            Common\FormatChangelogListener::class,
+            Version\PushTagToRemoteListener::class,
+            Version\CreateReleaseNameListener::class,
+            Version\PushReleaseToProviderListener::class,
         ],
         Version\RemoveChangelogVersionEvent::class => [
             Config\ConfigListener::class,
